Moves access preferences fragment classes within AdminPreferencesFragment class
