--- collect_app/src/main/java/org/odk/collect/android/fragments/MainMenuAccessPreferences.java
@@ -1,53 +1,0 @@
-/*
- * Copyright (C) 2017 Shobhit
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package org.odk.collect.android.fragments;
-
-import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.support.annotation.Nullable;
-import android.view.View;
-
-import org.odk.collect.android.R;
-import org.odk.collect.android.preferences.BasePreferenceFragment;
-
-import static android.content.Context.MODE_WORLD_READABLE;
-import static org.odk.collect.android.preferences.AdminPreferencesFragment.ADMIN_PREFERENCES;
-
-public class MainMenuAccessPreferences extends BasePreferenceFragment {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        PreferenceManager prefMgr = getPreferenceManager();
-        prefMgr.setSharedPreferencesName(ADMIN_PREFERENCES);
-        prefMgr.setSharedPreferencesMode(MODE_WORLD_READABLE);
-
-        addPreferencesFromResource(R.xml.main_menu_access_preferences);
-    }
-
-    @Override
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        toolbar.setTitle(R.string.main_menu_settings);
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        if (toolbar != null) {
-            toolbar.setTitle(R.string.admin_preferences);
-        }
-    }
-}
